# Restore trace callback when the profiler is disabled