--- a/docs/source/conf.py
+++ b/docs/source/conf.py
@@ -194,6 +194,9 @@ autosummary_mock_imports = [
     'geowatch.tasks.cold.export_change_map',
 ]
 
+autodoc_default_options = {  # Document callable classes
+    'special-members': '__call__'}
+
 autodoc_member_order = 'bysource'
 autoclass_content = 'both'
 # autodoc_mock_imports = ['torch', 'torchvision', 'visdom']