--- a/line_profiler/Python_wrapper.h
+++ b/line_profiler/Python_wrapper.h
@@ -18,6 +18,17 @@
     #include "frameobject.h"
 #endif
 
+#if PY_VERSION_HEX < 0x030900a4  // 3.9.0a4
+    #define PyObject_CallOneArg(func, arg) \
+        PyObject_CallFunctionObjArgs(func, arg, NULL)
+    #define PyObject_CallMethodOneArg(obj, name, arg) \
+        PyObject_CallMethodObjArgs(obj, name, arg, NULL)
+    #define PyObject_CallNoArgs(func) \
+        PyObject_CallFunctionObjArgs(func, NULL)
+    #define PyObject_CallMethodNoArgs(obj, name) \
+        PyObject_CallMethodObjArgs(obj, name, NULL)
+#endif
+
 #if PY_VERSION_HEX < 0x030900b1  // 3.9.0b1
     /*
      * Notes: