--- a/line_profiler/Python_wrapper.h
+++ b/line_profiler/Python_wrapper.h
@@ -5,45 +5,30 @@
 
 #include "Python.h"
 
-/*
- * Ensure PyFrameObject and PyInterpreterState availability as
- * concretely declared structs
- */
-#ifndef Py_BUILD_CORE
-    #define Py_BUILD_CORE 1
-#endif
+// Ensure PyFrameObject availability as a concretely declared struct
 
 // _frame -> PyFrameObject
 #if PY_VERSION_HEX >= 0x030b00a6  // 3.11.0a6
-    #include "internal/pycore_frame.h"
-    #include "cpython/code.h"
-    #include "pyframe.h"
+#   ifndef Py_BUILD_CORE
+#       define Py_BUILD_CORE 1
+#   endif
+#   include "internal/pycore_frame.h"
+#   include "cpython/code.h"
+#   include "pyframe.h"
 #else
-    #include "frameobject.h"
-#endif
-
-// _is -> PyInterpreterState
-#if PY_VERSION_HEX >= 0x030b00a6  // 3.11.0a6
-    #include "pytypedefs.h"
-#else
-    #include "pystate.h"
-#endif
-#if PY_VERSION_HEX >= 0x030900a6  // 3.9.0a6
-    #include "internal/pycore_interp.h"
-#else
-    #include "internal/pycore_pystate.h"
+#   include "frameobject.h"
 #endif
 
 // Backport of Python 3.9 caller hooks
 
 #if PY_VERSION_HEX < 0x030900a4  // 3.9.0a4
-    #define PyObject_CallOneArg(func, arg) \
+#   define PyObject_CallOneArg(func, arg) \
         PyObject_CallFunctionObjArgs(func, arg, NULL)
-    #define PyObject_CallMethodOneArg(obj, name, arg) \
+#   define PyObject_CallMethodOneArg(obj, name, arg) \
         PyObject_CallMethodObjArgs(obj, name, arg, NULL)
-    #define PyObject_CallNoArgs(func) \
+#   define PyObject_CallNoArgs(func) \
         PyObject_CallFunctionObjArgs(func, NULL)
-    #define PyObject_CallMethodNoArgs(obj, name) \
+#   define PyObject_CallMethodNoArgs(obj, name) \
         PyObject_CallMethodObjArgs(obj, name, NULL)
 #endif
 
@@ -54,7 +39,7 @@
      *     INCREF the code object until 0b1 (PR #19773), so override
      *     that for consistency.
      */
-    #define PyFrame_GetCode(x) PyFrame_GetCode_backport(x)
+#   define PyFrame_GetCode(x) PyFrame_GetCode_backport(x)
     inline PyCodeObject *PyFrame_GetCode_backport(PyFrameObject *frame)
     {
         PyCodeObject *code;
@@ -79,12 +64,12 @@
     {
         PyObject *code_bytes;
         if (code == NULL) return NULL;
-        #if PY_VERSION_HEX < 0x030b00a7  // 3.11.0a7
+#       if PY_VERSION_HEX < 0x030b00a7  // 3.11.0a7
             code_bytes = code->co_code;
             Py_XINCREF(code_bytes);
-        #else
+#       else
             code_bytes = PyObject_GetAttrString(code, "co_code");
-        #endif
+#       endif
         return code_bytes;
     }
 #endif