--- a/line_profiler/_line_profiler.pyx
+++ b/line_profiler/_line_profiler.pyx
@@ -202,7 +202,7 @@ def label(code):
     module in Python 2.5.
 
     Note:
-        In Python >= 3.11 we use we return qualname for ``_name``.
+        In Python >= 3.11 we return qualname for ``_name``.
         In older versions of Python we just return name.
     """
     if isinstance(code, str):
@@ -416,7 +416,7 @@ cdef class _LineProfilerManager:
         Calls :c:func:`legacy_trace_callback()`.  If
         :py:func:`sys.gettrace` returns this instance, replaces the
         default C-level trace function :c:func:`trace_trampoline` (see
-        the C implementation of :py:mod:sys`) with
+        the C implementation of :py:mod:`sys`) with
         :c:func:`legacy_trace_callback` to reduce overhead.
 
         Returns;
@@ -761,11 +761,12 @@ cdef class LineProfiler:
     cdef public object threaddata
 
     # These are shared between instances and threads
-    # type: dict[int, _LineProfilerManager], int = thread id
+    # type: ClassVar[dict[int, _LineProfilerManager]], int = thread id
     _managers = {}
-    # type: dict[bytes, int], bytes = bytecode
+    # type: ClassVar[dict[bytes, int]], bytes = bytecode
     _all_paddings = {}
-    # type: dict[int, weakref.WeakSet[LineProfiler]], int = func id
+    # type: ClassVar[dict[int, weakref.WeakSet[LineProfiler]]],
+    # int = func id
     _all_instances_by_funcs = {}
 
     def __init__(self, *functions,