(Closes #333.)

Summary
----

(**Update 12 Jul**: this summary has been completely rewritten to reflect the latest state of affairs)

This draft PR adds C(-ython) level support to:
- Cache the existing trace-callback state(s) when `.enable()`-ing line profiling and restoring said state(s) when `.disable()`-ing it
- Optionally call the cached callback(s) from within the profiler callback(s)
- Improve compatibility with code (e.g. `doctest.DocTestRunner.run()`) using this Python-level paradigm to temporarily alter "legacy" tracing:
  ```python
  callback = sys.gettrace()
  try:
      ...  # Do something which involves swapping/unsetting the trace function
  finally:
      sys.settrace(callback)  # Restore the trace function
  ```

The option to switch back to using the "legacy" system in Python 3.12+ is also added for debugging purposes.

### Code changes

<details>
<summary> (Click to expand) </summary>

- `line_profiler/Python_wrapper.h`  
  New C header wrapping `Python.h`, centralizing backports for the Python C API
- `line_profiler/c_trace_callbacks.{c,h}`  
  New C functions for handling interactions with the trace system:
  - `alloc_callback()`, `free_callback()`, `populate_callback()`, `nullify_callback()`, `restore_callback()`, `call_callback()`  
    Manage the retrieval, use, and restoration of C-level legacy trace callbacks (see `_LineProfilerManager`)
  - `set_local_trace()`  
    Set the frame-local legacy trace function of a frame object to the `_LineProfilerManager`, or attach the manager thereto
  - `monitoring_restart_version()`  
    Get the `.last_restart_version` from the internal interpreter state; this is necessary for detecting calls to `sys.monitoring.restart_events()` 
- `line_profiler/_line_profiler.pyx`
  - `label()`, `_code_replace()`, `LineStats`  
    Updated docstrings
  - `disable_line_events()`  
    New helper function for wrapping a frame-local legacy trace callable so that it no longer receives line events
  - `legacy_trace_callback()`
    - Updated call signature; now expecting a `_LineProfilerManager` (`manager`) instead of a `set[LineProfiler]` (`instances`) as the first argument
    - Now calling the cached legacy trace callback if `manager.wrap_trace` is true
    - Now setting the frame-local legacy trace function if `manager.set_frame_local_trace` is true
  - `_SysMonitoringState`  
    New thread-local (private) helper object for managing the interactions with `sys.monitoring`
    - `enable()`  
      Method to replace and cache the `sys.monitoring` state (relevant callbacks, tool name, active events) upon start of profiling
    - `disable()` 
      Method to restore the cached `sys.monitoring` state upon end of profiling
    - `call_callback()`  
      C-level method to call the cached callbacks in a safe way (i.e. which wouldn't interfere with line profiling) and manage said cache
  - `_LineProfilerManager`  
    New thread-local (private) helper object superseding the set-of-`LineProfiler`s as the object set as the legacy trace-callback object, which retrieves and caches the trace-callback state it replaces when profiling starts and restores them when it ends:
    - With the "legacy" trace system, `PyThreadState.c_tracefunc` and `PyThreadState.c_traceobj` are cached; note that `sys.gettrace()` only gets the latter and misses the former, hence the need for C-level shenanigans
    - With the `sys.monitoring`-based system, the callbacks returned by `sys.monitoring.register_callback()` for the events `sys.monitoring.events.LINE`, `.PY_RETURN`, `.PY_YIELD`, `.RAISE`, and `.RERAISE` are cached

    Class members:
    - `mon_state`  
      `_SysMonitoringState` instance
    - `__call__()`  
      Method providing the same API as Python-level legacy trace functions, so that round-tripping the object (`sys.settrace(sys.gettrace())`) doesn't cause an error
    - `wrap_local_f_trace()`  
      Method for creating wrappers for frame-local legacy trace functions which call both the instance and the wrapped function
    - `handle_line_event()`, `handle_return_event()`, `handle_yield_event()`, `handle_raise_event()`, `handle_reraise_event()`  
      Methods to be supplied to `sys.monitoring.register_callback()` for handling the `sys.monitoring.LINE`, `.PY_RETURN`, `.PY_YIELD`, `.RAISE`, and `.RERAISE` events and optionally calling the corresponding wrapped/cached callbacks; supersede the previous `monitoring_line_event_callback()` and `monitoring_exit_frame_callback()`
      - The cached callbacks and event set are kept updated
      - If a wrapped and cached callback somehow disables tracing (e.g. by trying to unset/change the `sys.monitoring` callbacks with `sys.monitoring.register_callback()`, calling `sys.monitoring.set_events()` to disable line events, or returning `sys.monitoring.DISABLE` to disable events for a certain code location), those changes are isolated so that line profiling is uninterrupted, but said callbacks will no longer receive the relevant event (except when `DISABLE`-d code locations are re-enabled by calling `sys.monitoring.restart_events()`)
    - `_handle_enable_event()`, `_handle_disable_event()`  
      New methods for handling `LineProfiler.enable()` and `.disable()`, managing the interactions with existing states of the tracing/monitoring system
  - `LineProfiler`
    - `__doc__`  
      Extensively extended docstring
    - `tool_id`  
      New class attribute for the `sys.monitoring` tool ID to use
    - `_manager`  
      New thread-local "class attribute" (`_LineProfilerManager` instance) which manages `LineProfiler` instances and provides the trace callbacks, superseding the previous `._active_instances`
    - <a name="line-profiler-wrap_trace">`wrap_trace`</a>  
      New init argument and (thread-local) "class attribute" to control whether existing trace callbacks should be called by the profiling callbacks (true) or stashed away (false)
    - <a name="line-profiler-set_frame_local_trace">`set_frame_local_trace`</a>  
      New init argument and (thread-local) "class attribute" to control whether to set `._manager` as the frame-local legacy trace function, which helps line profiling resume ASAP after legacy tracing is interrupted by e.g. `sys.settrace()`
    - `add_function()`  
      Added caveat warning against direct use; end-users should use `.add_callable()` from the wrapper class `line_profiler.line_profiler.LineProfiler`
    - `enable()`, `disable()`  
      Deferred most of the implementations to `_LineProfilerManager`
- `line_profiler/_diagnostics.py`  
  Added the following environmental switches:
  - `WRAP_TRACE` (env. var.: `${LINE_PROFILER_WRAP_TRACE}`)  
    Boolean flag (default `False`); default value for [`LineProfiler.wrap_trace`](#line-profiler-wrap_trace)
  - `SET_FRAME_LOCAL_TRACE` (env. var.: `${LINE_PROFILER_SET_FRAME_LOCAL_TRACE}`)  
    Boolean flag (default `False`); default value for [`LineProfiler.set_frame_local_trace`](#line-profiler-set_frame_local_trace)
  - `USE_LEGACY_TRACE` (env. var.: `${LINE_PROFILER_CORE}`; `old`, `legacy`, and `ctrace` resolve to `True`; `new`, `sys.monitoring`, and `sysmon` resolve to `False`)  
    Boolean flag (default `False` if `sys.monitoring` is available, `True` otherwise); toggles whether to use the legacy trace system or the new, `sys.monitoring`-based system introduced in #352

</details>

### Test changes

- `tests/test_line_profiler.py::test_sys_monitoring`  
  Updated implementation to always use `sys.monitoring` when available
- `tests/test_sys_monitoring.py`  
  New test module for the `sys.monitoring`-based system
  - `test_standalone_callback_usage()`, `test_wrapping_trace()`  
    Check that simple `sys.monitoring` callbacks behave as expected, and are invoked by `LineProfiler` when `wrap_trace=True`
  - `test_callback_switching()`  
   Checks that `sys.monitoring` callbacks which temper with `sys.monitoring.register_callback()` behave as expected, and are invoked by `LineProfiler` when `wrap_trace=True`
  - `test_callback_update_events()`  
   Checks that `sys.monitoring` callbacks which temper with `sys.monitoring.set_[local_]events()` behave as expected, and are invoked by `LineProfiler` when `wrap_trace=True`
  - `test_callback_toggle_local_events()`  
   Checks that `sys.monitoring` callbacks which return `sys.monitoring.DISABLE` and temper with `sys.monitoring.restart_events()` behave as expected, and are invoked by `LineProfiler` when `wrap_trace=True`
- `tests/test_sys_trace.py`  
  New test module for (mostly) the legacy trace system
  - `test_callback_preservation()`  
    Checks the restoration of the trace-callback state after using the profiler
  - `test_callback_wrapping()`  
    Checks trace-callback interoperability – that the existing trace callback can be used by the profiler's callback; also checks the interoperability of `LineProfiler.wrap_trace` and `LineProfiler.set_frame_local_trace`
  - `test_wrapping_throwing_callback()`  
    Checks the continuation of line profiling and the disabling of the existing callback in the event that it errors out
  - `test_wrapping_line_event_disabling_callback()`  
    Checks the continuation of line profiling and the hiding of line events from the existing frame-local callback in the event that it disables frame line events
  - `test_wrapping_thread_local_callbacks()`  
    Checks the usage and restoration of thread-local existing trace callbacks by the profiler
  - `test_python_level_trace_manipulation()`  
    Checks that line profiling is amenable to the `sys.settrace(sys.gettrace())` paradigm

### Packaging changes

- `docs/`  
  Now including the `.__call__()` methods of objects by default
- `line_profiler/CMakeLists.txt`  
  Updated list of C source files

Motivation
----

As noted in #333, `LineProfiler` currently interacts destructively with `sys.settrace()` and `sys.monitoring.register_callback()`, setting the corresponding trace callbacks to `None` when the profiler is `.disable()`-ed. Beside being in a way "not clean", this also has multiple unfortunate side effects, such as prohibiting cooperative use with other trace-callback based monitoring tools, such as coverage tools and debuggers.

Notably, the discarding of the trace callback was what tanked the coverage that `codecov` reported, because `coverage.py` could't see anything after any test called `LineProfiler.enable()` in-process. While #352 eliminated much of the issues with `codecov`, because `coverage.py` defaults to the legacy trace system while `line_profiler` defaults to `sys.monitoring` where available, it is still desirable for the tool to function correctly (by restoring existing callbacks) and cooperatively (by optionally enabling calling both the line-profiling and existing callbacks) with other tools.

Effects
----

### Python 3.8

| Run on\\Filename | Coverage (%, # lines, # missing) | Test duration/s (best, mean of 3) | Test duration/s (w/coverage) |
| :-: | :-: | :-: | :-: |
| `main` | 70, 1618, 425 | 11.99, 12.03 | 19.76 |
| This PR (as of 224d95d) | 78, 1618, 301 | 11.94, 12.01 | 19.87 |
| This PR w/`LINE_PROFILER_WRAP_TRACE=1` | 79, 1618, 288 | - | 20.25 |

### Python 3.13

| Run on\\Filename | Test duration/s best, mean of 3) | Test duration/s (w/coverage) | Test duration/s (w/coverage, `COVERAGE_CORE=sysmon`) |
| :-: | :-: | :-: | :-: |
| `main` | 14.55, 14.58 | 30.64 | 30.16 |
| This PR (as of 224d95d) | 14.72, 14.82 | 30.85 | 30.10 |
| This PR w/`LINE_PROFILER_WRAP_TRACE=1` | - | 30.89 | 29.97 |

### Notes

- Tests are run with the flag `-k "not test_cython and not test_sys"`, to (1) mitigate Cython-related issues with in some configurations and (2) ensure that the same set of tests is run between the different branches.
- The PR does not change coverage in Python 3.13 because `coverage.py` uses legacy tracing and functions independently of the rest of `sys.monitoring` (which `line_profiler` now uses). Also see comments below ([comment 1](https://github.com/pyutils/line_profiler/pull/334#issuecomment-3062404388), [comment 2](https://github.com/pyutils/line_profiler/pull/334#issuecomment-3065385285)).

Caveats
----

<details>
<summary> (Outdated; click to expand) </summary>

**(UPDATE 17 Apr: the optimizations mentioned below have been made.)**

As shown by the test-suite output above, this PR may introduce some overhead to the tracing function. However:
- The impact on the measured line timings should be minimal, since any interaction with the cached callback is after the collection of timing data.
- A considerable part of the apparent overhead may just have to do with how the cached callback is run more often (i.e. not being outright discarded) after this PR.

Running the entire test suite without `pytest-cov` shows the following results:

| Run on\\Real time elapsed (s; between 10 runs) | Mean | Best | σ |
| :-: | :-: | :-: | :-: |
| `main` | 7.62 | 7.42 | 0.180 |
| This PR | 7.65 | 7.42 | 0.210 |
| This PR w/`LINE_PROFILE_WRAP_TRACE=1` | 7.47 | 7.34 | 0.0983 |

Hence, in the case where we aren't interacting with other tracing facilities, the overhead is negligible and well within normal variations. Still, it may be worth looking into optimizing `wrap_trace` (as implemented in `line_profiler/_line_profiler.pyx::call_c_trace_state_hook_safe()`) since that seem to have more severe performance implications. Since it mostly deals with objects available on the C-level (`CTraceState`, `PyFrameObject.f_trace_lines`, `PyFrameObject.f_trace`), it's likely that  we can eke out more performance by moving `call_c_trace_state_hook_safe()` from Cython to C – or just merge it with the C-level `call_c_trace_state_hook()` function which it calls.

</details>

Update 17 Apr
----

<details>
<summary> (Click to expand) </summary>

### Reorganization

The tests which have to do with `sys` trace callbacks have been moved from the `tests/test_line_profiler.py` module to the new `tests/test_sys_trace.py`.

### Handling of edge cases

In order for trace-callback wrapping to be more transparent (i.e. for the wrapped callback to behave more as if it was free-standing): 
- In normal execution, a trace callback can unset the `sys` callback, either via erroring out or explicitly calling `sys.settrace(None)` (or its C equivalent). If a wrapped callback does so, e.g. in [`coverage.py::coverage/ctracer/tracer.c::CTracer_dealloc()` (L87)](https://github.com/nedbat/coveragepy/blob/287d0f9aedcc72e54d91a84d09e38e87e1b2a43b/coverage/ctracer/tracer.c#L87)):
  - The existing `sys` callback is restored so that profiling continues, as was previously done in this PR; but
  - The profiler's callback drops the cached callback from that point on, effectively disabling the latter; and
  - When the profiler is `.disable()`-ed, it unsets the `sys` trace callback instead of restoring it to the (dropped) cached value.
- In normal execution, a trace callback can set `.f_trace_lines` to false in a frame, disabling line events for the frame (which obstructs further line profiling). If a wrapped callback does so; e.g. in [`coverage.py::coverage/ctracer/tracer.c::CTracer_handle_call()` (L548)](https://github.com/nedbat/coveragepy/blob/287d0f9aedcc72e54d91a84d09e38e87e1b2a43b/coverage/ctracer/tracer.c#L548)):
  - `.f_trace_lines` is reset so that profiling continues, as was previously done in this PR; but
  - The frame's local trace callable `.f_trace` is wrapped so that it is no longer invoked for line events.

Moreover, support for multithreaded environments (where each thread can have its own `sys` trace function) has been added.

### Performance impacts

The previous timing info in "Effects" was unduly biased against the PR in that:
- It was later noted that the virtual environments hosting the `main` and PR branches had different dependency versions (e.g. `pytest`); updating them to be consistent substantially closed the gap.
- Since the new tests temper with the `sys` trace callback, they have to be run in subprocesses to isolate their side effects, resulting in relatively higher overhead.

As such, the venv dependencies have been updated to be consistent (`pytest==8.3.5`, `pytest-cov==6.1.1`, `coverage==7.8.0`), and we now run `run_test.py` or `pytest` with `-k 'not sys_trace'` to deselect the new tests. The updated timings are:

<table>

<tr>
<th>

With `pytest-cov` (`run_tests.py -k 'not sys_trace'`)

</th>
<th>

Without `pytest-cov` (`pytest -k 'not sys_trace'`)

</th>
</tr>

<tr>
<td>

| Run on\\Real time elapsed (s; between 3 runs) | Mean | Best | σ |
| :-: | :-: | :-: | :-: |
| `main` | 15.83 | 15.77 | 0.049 |
| This PR | 15.87 | 15.80 | 0.052 |
| This PR w/`LINE_PROFILE_WRAP_TRACE=1` | 15.83 | 15.80 | 0.033 |

</td>
<td>

| Run on\\Real time elapsed (s; between 10 runs) | Mean | Best | σ |
| :-: | :-: | :-: | :-: |
| `main` | 7.045 | 6.97 | 0.058 |
| This PR | 7.043 | 6.94 | 0.109 |
| This PR w/`LINE_PROFILE_WRAP_TRACE=1` | 7.159 | 7.0 | 0.101 |

</td>
</tr>
</table>

Hence the overhead introduced by the PR seems negligible. (The coverage improvements remain the same as in "Effects" above.)

</details>

Update 19 May
----

<details>
<summary> (Click to expand) </summary>

In response to review, the C-level functions in `line_profiler/_line_profiler.pyx` for handling trace callbacks are refactored out into their own C source/header files `line_profiler/c_trace_callback.c` and `.h`.

To make Python C-API use more consistent (e.g. back-porting of newer utils), the wrapper `line_profiler/Python_wrapper.h` is created around `Python.h`. As such:
- The need to refer to `PY_VERSION_HEX` in C(-ython) code elsewhere is minimized.
- A corner case which pertains to the use of `PyCode_GetCode()` in pre-release versions of Python 3.11 is caught and fixed.

</details>

<a name="section-update-26-may"></a>Update 26 May
----

<details>
<summary> (Click to expand) </summary>

The PR is completely rewritten after #347 was merged due to rebasing difficulties, but most of the previous changes are kept, with the following additions and enhancements:

- `line_profiler/_line_profiler.pyx`:
  - A new object type `_LineProfilerManager` is introduced so that:
    - The trace-callback state is decoupled from individual `LineProfiler` instances.
    - The object returned from `sys.gettrace()` is now:
      - An instance thereof (instead of being a `set[LineProfiler]`)
      - Callable as a Python-level trace function which returns itself, and
      - Amenable to being set via `sys.settrace()`;

     This improves compatibility when profiling code which does something like the below example (e.g. [`doctest.DocTestRunner.run()`](https://github.com/python/cpython/blob/e483dcf19eee7143ba76d1829afc1052053a904e/Lib/doctest.py#L1565)):
      ```python
      callback = sys.gettrace()  # This returns the manager object
      try:
          sys.settrace(something_else)  # This suspends line profiling
          ...
      finally:
          # This restores the manager object and resumes profiling,
          # but causes problems if it isn't callable w/the signature
          # `(frame: types.FrameType,
          #   event: Literal['call', 'line', 'return', 'exception', 'opcode'],
          #   arg: object)`
          sys.settrace(callback)
      ```
    The overhead should be minimal since the `.__call__()` resets the C-level trace callback to `python_trace_callback()` where possible, sidestepping `.__call__()` and the [default `sys` trace trampoline](https://github.com/python/cpython/blob/cf8941c60356acdd00055e5583a2d64761c34af4/Python/sysmodule.c#L1101) which calls it indirectly.
  - `LineProfiler.wrap_trace` is now a pseudo class attribute and synced between all instances.
…