@Erotemic 2025-04-14
Because this modifies C code, it's going to take more time for me to carefully review. I think this will be independent of the other changes you were interested in working on, but let me know if any PR is a blocker for other features. The feature I'm most interested in right now, is the one where `-p` will mark a submodule to be autoprofiled on import (or we could do it the easy way and just pre-import the specified modules in kernprof as an initial pass with plans to improve on it later).

With all of these changes, I'm also starting to think it would be a good idea for me to spend some effort building regression tests that monitor the overhead of line profiler on benchmark use-cases. Right now I currently test it out and check that it doesn't feel sluggish, but that's not sustainable.