@TTsangSC 2025-04-14
Fair points, please take your time.

Since we already have provisions for running "setup" code in `kernprof`, it's probably easy to also hack that for doing pre-imports. I'll give that a try, as well as try to port the AST stuff over in the longer run. Oh and of course the TOML stuff that we talked about yesterday.