@TTsangSC 2025-05-31
Hi, just in case you wanted to look at something else this is also ready and should have minimal conflicts with the other PRs (esp. #337). But the bad news is that since #347 made substantial changes to `line_profiler/_line_profiler.pyx` quite a bit had to be reworked (see [this section](https://github.com/pyutils/line_profiler/pull/334#section-update-26-may) in the PR description).