@Erotemic 2025-07-12
> We may want to think about whether the env vars ${LINE_PROFILE_WRAP_TRACE} and ${LINE_PROFILE_SET_FRAME_LOCAL_TRACE} are named appropriately. E.g. do we want ${LINE_PROFILE_*} or ${LINE_PROFILER_*}? 

I agree. I think having `LINE_PROFILE=1` to activate the global profiler makes sense (I read it more like an action than a setting), but for everything else where it is controlling behavior using `LINE_PROFILER_xxx` seems better. We could even alias `LINE_PROFILE=1` to mean something like `LINE_PROFILER_ACTIVATE_GLOBAL_PROFILER=1` (or something maybe a bit more concise)

> Or maybe are we gonna throw the env vars out once 

Controlling configuration with pyproject.toml will only make sense in some cases. I generally thinking having environment variables for config options like cibuildwheel does it is generally a good idea. But we can move there slowly.

> We may also want to provide the option somewhere to switch back to the legacy trace system, like how coverage allows for the backend to be specified by ${COVERAGE_CORE}. But that of course can wait.

An environment variable in this PR might make sense for this. If something goes wrong with the 4.3 release it will be useful to be able to say does it work if you set `LINE_PROFILER_CORE=legacy` (or whatever the variable is). 

I will work on reviewing this over the weekend.