@TTsangSC 2025-07-13
Implemented fixes for the points you brought up above. Really appreciated the input, there's a lot of the C API that I am not that familiar with.

However, I just noticed that there's unfortunately some unintended interaction between `LineProfiler.wrap_trace` and `LineProfiler.set_frame_local_trace`: 

<details>
<summary> Why we have a recursion error (click to expand) </summary>

- Say we're using the legacy system, both params are set to true, and we have a Python-level callback set via `sys.settrace(some_python_callback)`. Now `some_python_callback()` returns itself for a `'call'` event of a function so that it receives events and does whatever it does inside the function's frame.
- When calling that function (`'call'` event), our `manager: _LineProfilerManager` thus:
  - Caches the active callback: `manager.legacy_callback: TraceCallback = { trace_trampoline, some_python_callback }` (in `_line_profiler.pyx::_LineProfilerManager._handle_enable_event()`),
  - Replaces the active callback with `{ legacy_trace_callback, manager }`, which then
  - Sets the frame-local trace callback `frame->f_trace` to `manager` (in `_line_profiler.pyx::legacy_trace_callback()` -> `c_trace_callbacks.c::set_local_trace()`), and
  - Calls the cached callback via `c_trace_callbacks.c::call_callback()`, which then calls `some_python_callback()` with the `'call'` event; since `some_python_callback()` returns itself, `sysmodule.c::trace_trampoline()` sets it instead as `frame->f_trace`...
  - ... but since we know that will happen and it defeats the purpose of `.set_frame_local_trace`, another call to `set_local_trace()` in `legacy_trace_callback()` sets the `frame->f_trace` to the wrapper `manager.wrap_local_f_trace(some_python_callback)`.

  So far so good – we've ensured that Python code that tempers with `sys.settrace()` and `sys.gettrace()` will not interrupt line-profiling of the frame.
- However, when we then proceed through the function body (`'line'` event):
  - `legacy_trace_callback(manager, ...)` is called, which then calls `call_callback()`
  - Since the cached callback uses `trace_trampoline()`, it calls `frame->f_trace`, which is now our wrapper around `some_python_callback()` instead of said callback itself...
  - ... but the wrapper also calls `manager()`, which calls `legacy_trace_callback()`, causing a recursion.

</details>

Honestly it's on me – I didn't think to test the two features side-by-side before, just individually. Will try to code up a fix ASAP (or maybe we just make it so that the two can't be used simultaneously).