@Erotemic 2025-07-18
I did another pass, and I think it looks good. That being said, its easy to miss something with a big refactor like this and additional C code. I'm going to merge, and then we just have one more PR before we can release. 

I do wonder if it would be possible to write a rust backend where we can prove there are no memory errors, and if that would be as performent as the current cython+c implementation. I think that would be a possible and useful AI task in the not so distant future.