@TTsangSC 2025-07-18
Again, thanks for the review!

This reminds me, I really should be learning Rust. Not entirely sure though if there would be real benefits in terms of memory safety (since at the end of the day we still have features requiring talking to the Python interpreter via the C API – with both the legacy and `sys.monitoring` backends), but I guess it wouldn't hurt.