@Erotemic 2025-07-18
There likely would be unsafe sections, but it would restrict any scope of concern to those sections, which is the main draw. I imagine it is only a matter of time before Python is re-written in Rust. It may be quite a bit of time, which means that now is a great time to be learning it.